Add a Go workspace
This is an abnormal usage of Go workspaces, instead
primarily for convenience.

This is due to ease of understanding of the examples in each chapter
code. For example, when a third-party package is added, it'll only show
up in the relevant chapter's module.

See the following to learn more about Go workspaces:
https://go.dev/doc/tutorial/workspaces